# perf(encoding/form): optimize camelCase and snakeCase conversion